+++ ChangeLog.md
@@ -1,3 +1,15 @@
+### 0.1.1 / 2016-08-15
+
+* Xzanth's modifications:
+  * Arguments now a list of hashes:
+    * name specifies name of argument
+    * format its format
+    * optional whether the argument is optional or not
+  * `!help` command can now respond through different means:
+    * notify
+    * query
+    * reply
+
 ### 0.1.0 / 2013-06-02
 
 * Initial release:
@@ -6,4 +18,3 @@
     * Usages
     * Summaries
     * Descriptions
-